chore: refactor genre handler folder
